refactor: extract triangulation module
