+++ crates/kornia-3d/src/pose/mod.rs
@@ -22,5 +22,8 @@ pub use essential::*;
 mod twoview;
 pub use twoview::*;
 
+mod triangulation;
+pub use triangulation::*;
+
 mod pose3d;
 pub use pose3d::*;
